refactor: 가격 형타입 변경
할인률, 쿠폰 적용으로 인한 부동 소수점 오차를 없애기 위해서 형타입 Long에서 BigDecimal로 변경

+++ src/main/java/com/bookstore/product/domain/Price.java
@@ -1,5 +1,6 @@
 package com.bookstore.product.domain;
 
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import lombok.Getter;
@@ -15,11 +16,11 @@ public class Price {
     private static final String PRICE_LESS_THAN_ZERO = "가격은 0원보다 작을 수 없습니다.";
 
     @Column(nullable = false)
-    private Long price;
+    private BigDecimal price;
 
     public Price(Long price) {
         validate(price);
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     private void validate(Long price) {
